refactor: config 패키지 domain 모듈로 이동
