+++ qrust-api/src/test/java/com/qrust/qrcode/infrastructure/minio/QrCodeMinioUploadImplTest.java
@@ -17,7 +17,6 @@ import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
 @Tag("minio")
-@ActiveProfiles("harvey")
 class QrCodeMinioUploadImplTest extends QrCodeTestTemplate {
 
     @Autowired
