+++ qrust-domain/src/main/java/com/qrust/common/config/FeignConfig.java
@@ -1,0 +1,10 @@
+package com.qrust.common.config;
+
+import com.qrust.external.google.infrastructure.GoogleSafeBrowsingFeignClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableFeignClients(clients = GoogleSafeBrowsingFeignClient.class)
+public class FeignConfig {
+}
